Disable version update in UpdatePluginPackage
The functionality to update the version attribute in the `UpdatePluginPackage.cs` file has been disabled by commenting out the relevant code. This decision was made based on the Dataverse documentation on Microsoft's official site, which states that the version attribute is not updatable. The original code, which attempted to either modify the existing version attribute or add it if absent, has been replaced with a comment block. This block includes a reference to the official documentation to justify the necessity of this change.